Refactor: move exceptions into a central place
